requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
cached-property==1.5.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
defusedxml==0.7.1
dill==0.3.7
dnspython==2.4.2
exchangelib==5.1.0
fastapi==0.103.2
fonttools==4.43.1
frozenlist==1.4.0
h11==0.14.0
hiredis==2.2.3
idna==3.4
isodate==0.6.1
isort==5.12.0
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
loguru==0.7.2
lxml==4.9.3
matplotlib==3.8.0
mccabe==0.7.0
multidict==6.0.4
numpy==1.26.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
platformdirs==3.11.0
pyaml==23.9.7
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.16.1
pylint==3.0.1
pyparsing==3.1.1
pyspnego==0.10.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
tomlkit==0.12.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
uvicorn==0.23.2
wrapt==1.15.0
yarl==1.9.2
